Fixed Cursor overriding HintText.
Since the introduction of a Cursor in version 3.2.0, if enabled said Cursor would override any hint text due to them being displayed in the same location on screen (both centered in the box).
This commit resolves the issue by painting the Cursor with an offset if a HintText is also provided.
Depending on the current TextDirection the Cursor is painted to the left (LTR) or right (RTL) of the HintText.
This mirrors the implementation of most similar cases e.g. the search field on StackOverflow.